Add `At` Lua draw commands for matrix transforms
Much of the work on the Lua side of things is due to that the draw API
expects coordinates in screen space while most of the data is defined
in tree space. That is manually transformed for each position and vertex
submitted from Lua for the `DrawImage` and `DrawImageQuad` commands.

Here we add corresponding `DrawImageAt` and `DrawImageQuadAt` commands
that take an additional transformation matrix, allowing the Lua program
to offload transformation work to the runtime by submitting geometry
in whichever space it wants with a 2x3 transformation matrix (actually a
3x3 matrix with an implied row of `0 0 1` for homogeneous transforms) to
bring it into screen coordinates.

As this untangles the transform from the geometry, it paves the way for
pre-allocated geometry meshes as much of the geometry in something like
the skill tree is static.