Add `Mesh` type for immutable 2D triangle meshes
Especially for the passive skill tree there is much geometry that is the
same across many frames and often the whole program lifetime which when
submitted with the existing functions is marshalled across the API
boundary over and over again, also taking up significant space in the
draw command buffer.

This change introduces a prepared `Mesh` type that represents an
immutable 2D triangle list of positions, texcoords and indices.
Instances of this type can be drawn with an image texture of choice with
a transform to place and size the mesh, extending the previous `At`
family of draw commands.

These handles greatly reduce the amount of work needed to validate
arguments in Lua draw calls. Currently as a proof-of-concept they are
processed into several degenerate quad draw calls as there is no
dedicated draw command for meshes in the renderer yet.