Add draw command for prepared meshes
This introduces a new `MESH` draw command for the layer queue which
encodes a handle to a prepared mesh and a transform to orient and scale
it.

This is significantly more compact than the previous choices of
decomposing the geometry into quads or triangles (as degenerate quads),
instead being able to prepare indexed geometry up-front and refer to it
by handle.

In order to store these meshes the renderer gains a list of mesh slots
that hold a pointer to the actual mesh object and a generation. Outside
of the renderer the mesh is referred to as a mesh handle consisting of
the index of the mesh slot and the generation.

Mesh destruction is deferred to the end of the frame so that handles can
be created, put in the queue and destroyed in the span of a single frame
without reference counting or copying.

The generation is essential for frame hashing as otherwise it's possible
to create another mesh in a recently vacated slot for the next frame in
the same draw order. The generation ensures that subsequent frames with
meshes occupying the same slots are still considered distinct. As an
effect of this it's not a good idea to recreate meshes every frame as
they will always be considered distinct even if they contain identical
geometry as last frame.