Make mesh API handle both triangles and quads
To make the use sites on the Lua side more succinct and not have to
submit as much data when making "regular" geometry the `NewMeshHandle`
function now has a new first parameter that indicates if the indices are
omitted, the vertex data can be interpreted as lists of triangles and
quads respectively.

Additionally, this allows quads to omit the texture coordinate array if
all the quads use the whole image. Triangles still require texcoords as
there's no sensible default for them and they're likely more complex
geometry to begin with anyway.

In order to support C++ RAII classes in the functions exposed in the API
to Lua this change also introduces a new assertion `LExpect` functioning
like the existing `LAssert` but instead of directly transferring control
to Lua it pushes the message on the Lua stack and throws a C++ exception
to unwind the C++ call stack. For this scheme the convenience macros
`SG_LUA_CPP_FUN_BEGIN(Name)` and `SG_LUA_CPP_FUN_END()` wrap the
function entrypoint and perform the second phase of assertion processing
when needed.