feat: yield UTF-8 paths from file searches
Previously the file search objects exposed to Lua accepted and generated
paths encoded in the user's narrow codepage. As a move toward UTF-8
awareness for paths and strings on the Lua side, we now instead accept
and generate UTF-8 strings.

This should be robust against legacy paths represented in the user's
codepage, the only source of such paths should be if they're
historically persisted in locations like `Settings.xml`.