feat: detect Wine for graphics API choice
We can detect Wine in a nice way to make compatibility choices for
things that cannot feasibly be fixed in upstream Wine, like picking an
ANGLE backend more native to the platform like Vulkan rather than a
Direct3D one.

Such things require future GLFW 3.4 so not immediately possible, but
this commit lays the groundwork even if it has to be moved elsewhere
eventually.