feat: remove UTF-8 impls for Lua io/os functions
As these functions can be fixed inside of the LuaJIT interpreter itself,
we don't need these hooked into the main interpreter anymore.

A positive side effect of this is that subscripts now also can use UTF-8
in paths as the changes are universal across all LuaJIT interpreters.

The fallback regular Lua interpreter in `Update.exe` is still narrow ACP
and care needs to be made when writing update op-files and modifying
`UpdateApply.lua`