feat: working text tabs, colour, cursor mapping
Tabs are as wide as the font size is tall as that maps closely to the
old font rendering.

Text colour is implemented by carefully keeping track of where colour
transitions in the rich text strings and applying them at the correct
time while rendering.

Cursor mapping is done by walking the horizontal extents of each glyph
in the layout. Tabs are special-cased and will position the cursor
either before or after the tab based on where on the void the cursor
query happens.