fix: install DLLs for dependencies on Windows
Some kinds of targets from vcpkg haven't considered the deployment story
and do not contribute any DLLs to the install target. To address this
we hand-find the libraries in the deps structure and install them into
the runtime directory. This is needed in CI to include the libraries
into the deps archive.